refactor: move tokenizeSearchQuery function to a new utility file and add tests for its functionality
