+++ apps/studio/src/schemas/resource.ts
@@ -109,10 +109,6 @@ export interface ResourceItemContent {
   parentId: string | null
 }
 
-export const tokenizeSearchQuery = (query: string): string[] => [
-  ...new Set(query.trim().toLowerCase().split(/\s+/).filter(Boolean)),
-]
-
 export const searchSchema = z
   .object({
     siteId: z.string(),
