+++ apps/studio/src/server/modules/resource/__tests__/resource.utils.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest"
 
-import { tokenizeSearchQuery } from "../resource"
+import { tokenizeSearchQuery } from "../resource.utils"
 
 describe("tokenizeSearchQuery", () => {
   it("returns empty array for empty string", () => {
