+++ apps/studio/src/server/modules/resource/resource.utils.ts
@@ -1,0 +1,3 @@
+export const tokenizeSearchQuery = (query: string): string[] => [
+  ...new Set(query.trim().toLowerCase().split(/\s+/).filter(Boolean)),
+]
